Added processing of Long Poll responses for VK - Now all changes occurring in the VK are displayed instantly Change in the mailing list: - Now, when the mailing list starts, the list of selected dialogs is cleared

+++ app/src/main/java/com/progcorp/unitedmessengers/interfaces/requests/VKConversationByIdRequest.kt
@@ -9,7 +9,7 @@ interface VKConversationByIdRequest {
     suspend fun conversationGet(
         @Query("access_token") token: String,
         @Query("v") v: String,
-        @Query("peer_ids") peer_ids: Int,
+        @Query("peer_ids") peer_ids: Long,
         @Query("extended") extended: Boolean,
         @Query("lang") lang: Int
     ): Response<String>
